Made attributes and notifToXml() private
