+++ flutter_local_notifications_windows/lib/src/details.dart
@@ -9,7 +9,6 @@ export "details/notification_input.dart";
 export "details/notification_part.dart";
 export "details/notification_progress.dart";
 export "details/notification_text.dart";
-export "details/notification_to_xml.dart";
 
 /// The result of updating a notification.
 enum NotificationUpdateResult {
